refactor: canonical vault resolver in @polpo-ai/core
resolveAgentVault/ResolvedVault lived in the root polpo-ai package
(src/vault/resolver.ts) and were not exported from any scoped package,
so consumers kept local copies in sync by hand: a looser duplicate
interface in @polpo-ai/tools and three inline copies in the cloud data
plane (which cites the missing export as the reason).

The resolver moves to @polpo-ai/core as pure logic (environment is
injectable; defaults to the host env when one exists). The shell
re-exports it (shim) and tools re-exports the canonical type — every
consumer now shares one definition, and the cloud can drop its copies
at the next version bump.
